fix(brain): KR 进度采集链路修复 — verifier $2 类型冲突 + 每日健康审计 + migration 编号去重
修复 KR2 verifier 持续报错 "inconsistent types deduced for parameter $2"（UPDATE key_results 中 $2::text 与
numeric 列类型推断冲突），改为 ($2::numeric)::text 明确类型。

新增 kr-daily-audit.js：每天自动汇总所有 KR verifier 健康状态，计算 trust_score，
写入 event_log，并通过 GET /api/brain/okr/health 暴露实时查询。

修复 migration 编号冲突：224_fix_kr34_progress_verifiers.sql → 225_*，
更新 selfcheck.js EXPECTED_SCHEMA_VERSION 和 DEFINITION.md。

+++ packages/brain/src/kr-verifier.js
@@ -71,7 +71,7 @@ export async function runAllVerifiers() {
           UPDATE key_results
           SET progress = $1,
               current_value = $2,
-              metadata = COALESCE(metadata, '{}'::jsonb) || jsonb_build_object('metric_current', $2::text),
+              metadata = COALESCE(metadata, '{}'::jsonb) || jsonb_build_object('metric_current', ($2::numeric)::text),
               updated_at = NOW()
           WHERE id = $3
         `, [progress, currentValue, v.kr_id]);
